pre_screening_livekit_agent/requirements.txt
livekit-agents[openai]~=1.4
livekit-plugins-noise-cancellation~=0.2
python-dotenv
